GIR: Write and read allow-none for property values
allow-none for properties is not strictly a GIR attribute. However, in
the light of writing libraries with vala and using only GIR I think it would
be good to add this property attribute. Note that GIR format does
allow custom attributes to be set on any element so it's not a
violation.

https://bugzilla.gnome.org/show_bug.cgi?id=669640